Initial implementation of Odyssey code update wrapper procedure
This defines some data structures and the HWP that can be used
to drive the code update of the Odyssey firmware image(s) and
get the current code levels.

This defines the subroutine calls for performing a code update
and getting the current levels.

Note - Implementation is incomplete, here for initial discussion
and expansion as more details are finalized.

Change-Id: I28f93f6ea5f71be345730bafc9a31743699f0167
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/145340
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Dev-Ready: Steven B Janssen <janssens@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Brent Wieman <bwieman@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149875
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>