Tweaks for HBRT Testcase
Changes the loader.H HBRT test to use mm_alloc_block instead of malloc
which was causing out-of-memory task crashes since HBRT requires many
contiguous pages for malloc. In testing, available memory is already
under heavy use so it's not always possible for this malloc to succeed.
By putting HBRT into the VMM space it resolves the issue.

Change-Id: Ia6746b120e4602b192a1d0a4334424eca35b7455
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/145460
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>