eBMC setup standalone script: fix cherry-picking and MACHINE checks
Adds a cherry-picking strategy to prefer the incoming changes of the
standalone patching commit to avoid merge conflicts with the existing
P10 standalone pnor layout. Also, checks that the MACHINE env var is
setup properly because if it's not that will be a source of quiet pain.

Finally, removed the NIMBUS DEFAULT_MACHINE env var.

Change-Id: Ia10e21e47c7be681c1a1e6fdcff4fc26904bae1e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/146686
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>