Increate size of Huge allocation slots to 3MB
We discovered there are a few paths in our testcases that allocate
more than 2MB with a single malloc.  This is HBRT code so we can't
use explicit mm_alloc_block calls.  As currently written this
large size will exceed the Huge allocation range that we have (1MB)
so will revert back to a normal Big allocation that requires
contiguous pages.

Change-Id: Ic1039693452fd0dd28c2f1c02b6aa4b64895d784
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/146763
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Nick Bofferding <bofferdn@us.ibm.com>