Enable Physical Hierarchy Lookups
For consumers of tryGetAttributeInHierarchy, the use cases are looking
up attributes related to FRU components.  Ensure that the physical
path is the information being retrieved to properly add FRU callout
information as needed.

The caller of the tryGetAttributeInHierarchy needs to specify
if the desired requirement is AFFINITY or PHYSICAL for
the desired parentage lookup.

Change-Id: I360c9117496d863eeccd46529a7894b443e8b40e
JIRA: PFHB-11
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/146945
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>