IO: Turn const by-reference integer parameters into by-value
Used in a function parameter context, a constant reference to an integer
is functionally equivalent to a simple by-value parameter, but the code that
is generated ends up being smaller and simpler. Most of all though, this
style of parameter passing hit a PPE compiler bug :)

Replace all occurrences in the IO HWPs by by-value parameters to fix this.

Change-Id: I69f7e3ca55d8a8c21025f99080b6b7daec6bd97a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147658
Reviewed-by: Chris Steffen <cwsteffen@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147666
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>