DCE: Support for standalone simics
This commit makes several changes to support DCE in standalone
simulation:

- hb-attrWrite simics command allows setting attributes on the fly
  (can be run any time, writes are picked up between isteps)
- hb-enableDCE enables DCE in simulation (must be run before boot)
- hb-pauseIstepsAt uses the attribute write commands to pause
  isteps at a given point (can be run any time)
- hb-executeDCELid executes a DCE program (any time)
- hb-simicsLPCConsole redirects the LPC console to the SIMICS console
  (can be run any time as well)

DCE must be used with the test image (hb prime --test). It will not
function with non-test images (because libpldm_extended is not loaded
in non-test standalone images). DCE binaries should be built using the
command `make <name>.dce.test.lid` instead of the usual
`make <name>.dce.lid` command.

Running hb-enableDCE sets a flag in the debug framework which causes
Hostboot to start a monitor thread. This monitor thread checks
periodically whether the user has requested to execute a DCE program
and if so, downloads and runs the binary. There is no overhead in
non-simulation environments or when the hb-enableDCE command is not
run. (This command is run automatically in standalone simulation.)

hb-attrWrite can be run at any time, and new writes are applied
at the beginning of each Istep.

hb-executeDCELid can be run at any time after the extended image is
loaded.

The hb-enableDCE command must be run early in the IPL, preferably
before even allowing the simulator to run. (It is run automatically in
standalone simulation.)

A typical use case would be to use hb-pauseIstepsAt to ask Hostboot to
wait at the desired istep, then use hb-executeDCELid to iterate on a
piece of code.

This support is also compatible with simics checkpoints. Set up DCE
once and save/reload a checkpoint to avoid paying the setup cost each
time you restart the simulation.

Change-Id: Idce2dce2dce2dce2dce2dce2dce2dce2dce2dce2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147926
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>