Add SBE FIFO command class and command to an interface check
Fixes a compile issue where the sbeioInterfaceChecks in
getFifoSbeCapabilities was missing the command class and command of the
SBE FIFO message in the function call.

This managed to go unnoticed in review as the commit that changed the
function parameters for sbeioInterfaceChecks was added after CI had
already passed on the commit where getFifoSbeCapabilities was not
updated to match the new parameters. In the commit that changed these
parameters, getFifoSbeCapabilities had not yet added the call to
sbeioInterfaceChecks.

Change-Id: I3806edbc529db33ac1663616f7bc3bce9bc44a1f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147970
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>