ebmc setup script: Fix MACHINE env var checks
Updates the env var MACHINE checks to look for an exact match on the
search for that var. Some environments may have other *MACHINE* vars set
which can cause an erroneous failure. Additionally, since the MACHINE
var is case insensitive when used in the various hostboot util scripts
this script also allows for that var to be case insensitive when
performing the comparison to the expected value.

Change-Id: Iae7a9907fc7a67487bc1a236c7d1a17aecf4c5e6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147975
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>