DCE: Change DCE file suffixes to C/H from c++/h++
This commit transitions DCE file suffixes over to C/H instead of
c++/h++. This reduces the amount of effort required to setup, test, and
make changes to multiple files that are to be committed to the Hostboot
repository or already existed in the Hostboot repo.

Additionally, this commit includes tracinterface.H and trace_defs.H in
the compile of the DCE lid so that TRACFCOMP is redirected to output to
the console. That way, by default, TRACFCOMP calls inside of DCE files
automatically output to the console without a user needing to put forth
additional effort to do so. As a result, other macros that forwarded to
TRACFCOMP are also output to the console by default as well.

Change-Id: I0e92f5848e4fa4d7e18dba6916ff0b56c88daf27
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148086
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>