Fix isdimm presence detection override bug
The logic added to support DDR5 logical dimm and memport states
broke the scenario where an ISDIMM was unplugged.  The logic
would end up enabling the missing DIMM target after it was
already found to be not present.  The fix is to only downgrade
the DIMM's state, never upgrade, by only applying state changes
if the DIMM was found to be present initially.

Change-Id: Ia88f474c596ed55b3a6fa2e995a1b3f0250bd33a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148193
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>