PRD: masking PLL attentions on non-functional units
The test team was able to inject a PLL attention during istep 6 on an MC
that either is, or will be, deconfigured. The first chance PRD has to
analyze the attention is istep 9.1, at which point the MC unit is
definitely deconfigured. However, the PLLs are still active for all
chiplets. Normally, PRD will only iterate on functional targets, but
that does not work in this scenario because the underlying PLL attention
on the deconfigured MC won't be masked. Instead we will have to bypass
the PRD framework and manually mask all PLLs.

This workaround is only changes the behavior of masking PLL attentions.
This is not needed for detection nor clearing because broadcast register
is used to determine a PLL attention exists and also used to clear the
attentions. Unfortunately, FFFC collection will only capture registers
from functional units. There really isn't a way around that without
creating PRD objects for non-functional chips, which would have rippling
effects throughout all of PRD.

Change-Id: Ia20f19c995804a1803b2bebe782fd89ba541a82a
STGD: 411376
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148200
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148266
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148375
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>