Add support for Odyssey FFDC packages to SBE FFDC Parser
Adds support for Odyssey FFDC packages sent by the SBE to Hostboots SBE
FFDC package parser. Odyssey FFDC packages have an extra word in the
header and a unique magic byte that needed to be accounted for during
parsing.

Additionally, this commit fixes a memory leak and integer overflow that
was present in the parser class. As well as eliminating the multiple
redefinitions of the const for maximum page size which FFDC buffers can be.

Change-Id: I9f593f318da4efeb0cac08c9384ac1366ab97e6f
JIRA: PFHB-260
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148823
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>