Odyssey Fixes for Hostboot
This commit includes a few fixes that I had to do as part of Istep11
standalone Odyssey work in Hostboot.

The fixes are:
- Chip ID needs to include the location code
- The ody_sppe_check_for_ready poll timeout needs to be increased,
since none of the OCMBs come up within the current timeout on P10
standalone (the slowest took 55 polls)
- Some attributes in Odyssey HWPs need to have fapi2:: prefix, otherwise
they won't compile in Hostboot. Adding the prefix to the attribute names
was the cheapest fix in this case.

Change-Id: I0cf8952fc1f233a3c672a0e60a9625a04f3adfdd
JIRA: PFHB-411
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148717
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148741
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>