Method to set and remember the side the SPPE boots on
Add a new attribute to be able to specify the side to boot
the SPPE on.  Allows later HWP to utilize this attribute to
determine actions based on how it was booted.

Update mod_cbs_start to read the attribute and set the
appropriate register on the SPPE

Change-Id: If9b2d61198b62e7ffcaef8c7c4abb5996e236513
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148936
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Dev-Ready: Brent Wieman <bwieman@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/148948
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>