MRW processing support for doubled logical dimms in DDR5
The Odyssey DDR5 OCMB supports two memory ports, each with its
own logical DIMM target.  This presents some problems because
up to this point there has been an implicit 1-to-1 relationship
between the physicall DDIMM instance in the MRW and the logical
DIMM target in Hostboot.  To address this, processMrw.pl will
now key off of the "ddr" instances below the DDIMM in the MRW.

   <targetInstance>
	<id>ddimm-0</id>
	<type>lcard-dimm-ddimm</type>
	<is_root>false</is_root>
	<instance_name>ddimm</instance_name>
	<position>0</position>
	<child_id>generic-logic-assoc</child_id>
	<child_id>ocmb</child_id>
	<child_id>pmic0</child_id>
	<child_id>pmic1</child_id>
	<child_id>spdA</child_id>
	<child_id>spdB</child_id>
	<child_id>ddrA</child_id>   THIS
	<child_id>ddrB</child_id>   THIS

The DDIMM instance contains a lot of useful information so
it is still processed, but it is eventually removed from the
list of targets that are part of the final output.

As part of these changes, support was added to handle DDR4, DDR5,
and generic DDR instance and bus types.  This helps to maintain
backward compatibility for existing MRWs as well as allow more
accurate content going forward as systems support both DDR4
and DDR5 DDIMMs at the same time.

A tweak to the PDR handling was required as part of the DIMM
numbering changes.  The two logical DIMM targets that are part
of the same physical DDIMM will share an instance number.

During test and review problems were discovered in the simics
xml files.  These numbering issues on the dimm targets have
also been corrected.

CMVC-Coreq: 1167916
Change-Id: I31554991fa9e7b07636bfc2b35eda2f72d94cb0c
JIRA: PFHB-250
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149032
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>