PRD: fix trailing newline on auto-gen JSON files
The JSON output from our python files will not add a trailing newline on
the last line of the file. Most editors, especially vim and VS Code,
will automatically add this new line when it saves the file. This causes
annoying diffs and some mirror errors. I updated all of our python tools
that dump JSON data to file to add a newline to the end of the file.

Change-Id: Iecd4d519561784a3ffe12576f6a0702cb6dc7ba8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149050
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149214
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>