Remove PHYP (POWERVM) from HB Reserved Memory
PHYP no longer requires the pre-verified POWERVM lids
to be registered in the Hostboot Reserved Memory data
structures.

When the compression was added for the POWERVM lids,
the sizes of the compressed lids versus the sizes of
the uncompressed lids differ.  To remove the data
handling nuances the design decision was to have PHYP
no longer cross verify with the Hostboot Reserved
Memory structures for the POWERVM lid content which
is uncompressed directly to the PHYP mainstore memory.

Change-Id: I63caa0b4db42077663a950e16da7817423a0d86a
STGD: 381949
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149552
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>