Refactor hdatLoadMsData
This commit refactors the function hdatLoadMsData to call a new helper
function in an effort to reduce the extreme indenting that was present
in the function.

This has been done ahead of some required changes to support Odyssey in
this function. There have been no functional changes to the code in this
commit beyond fixing a memory leak.

Change-Id: Ieef0642012fed0fca370e5685798a8c2746268ad
JIRA: PFHB-387
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149760
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>