Add attributes for reallocating power from DIMMs
The power requirements for some DIMMs can be reduced based on their
physical location. The new PREHEAT_PERCENT will indicate the power
percentage that impacts the processor. It will be consumed by TMGT
and power relayed to the OCCs. TMGT will write the DIMM_POWER attribute
to an OCMB target after taking into account the preheat for all DIMMs
under that OCMB.

Change-Id: I3cb54d7b6fcb82e393765b8e605e5b5ddf256219
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150431
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Martha Broyles <mbroyles@us.ibm.com>
Reviewed-by: Brian T Vanderpool <vanderp@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>