Odyssey code update: Check for update needed
The Odyssey family of OCMBs support a chipop called "getCodeLevels"
which returns the hashes of the different binary images that it
contains in its non-volatile storage. Hostboot shall use the
information from this chipop to determine, among other things, whether
it needs to perform a code update on the OCMB.

This commit implements support for the getCodeLevels chipop, and uses
it to compare the code levels to what Hostboot has in its OCMBFW PNOR
partition to do the code update check.

Currently, Hostboot doesn't actually do anything with this
information, but in a future story it will actually initiate a code
update as appropriate.

Change-Id: I85a4502db016f545c15e88d0fc903b3cce040995
JIRA: PFHB-440
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149869
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>