Move CBS cleanup steps into poz_tp_init
After the CBS has been started no outside code may touch any of the mailbox
registers any more since it could (and will) clash with code running on the
SPPE.

Actual lab fail case: The cbs_cleanup code (outside) was read-modify-writing
ROOT_CTRL0 while the SPPE was executing the command table. This caused the
outside code to restore the value of the PCB_RESET bit (cleared by command
table code) to 1, causing SCOMs to fail and the SPPE to crash.

The fix: Move the cbs_cleanup steps to either _before_ the CBS is triggered,
or to a place where it runs either inside the SPPE or the SPPE is quiesced
if it runs in Cronus - namely poz_tp_init.

Change-Id: Ie85f4c5c0ad1d77026c1be17419c1c076b4a6795
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150742
Dev-Ready: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150753
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>