PM OCC FIR: make occ_complex_notify an XSTOP for debug Phantom Interrupt BCE
THIS COMMIT MUST BE UNDONE BEFORE GA!!!!!!!

In order to debug the OCC ASYNC_PHANTOM_INTERRUPT_BCE issues in DVT and CI
runs to OCC reset timeouts, OCC LFIR(58}[occ_complex_notify] was set to create
a checkstop.  In the field, though, this event will be recovered by the PM
Complex reset so the checkstop setting CANNOT go to the field.

Change-Id: I057e2d1f8683fd73b56a9c39f24eedfb2306a735
CCSD: 382710
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151739
Reviewed-by: DAVID DU <daviddu@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: DAVID A KING <kingboy@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151747
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>