Odyssey: Attr Sync - Pull
This commit adds logic to pull attribute blob from Odyssey and  apply
the values received in the blob to Host-side attributes. We receive
the attribute blob as a binary from Odyssey and use a HWP to parse
and apply the new values. This happens on all functional Odyssey
chips in the system in istep 13.2

Change-Id: I711c10efb82569dfb4fe2f2f67c2c5a6b1f3748b
JIRA: PFHB-490
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152713
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>