Hostboot Building Patches for eBMC in Standalone
Changes the setup-standalone-repo script to be more flexible to changes
introduced in 1060 while remaining backward compatible by doing the
following:

Copy OP-BUILD PNOR XML from the proper backing build
to allow the PNOR to be built in standalone.

Export BMC_STANDALONE_PATCHING=1 to indicate to
assemble.standalone.mk to model the PNOR appropriately.

Change the commit that was cherry-picked to a new one which modifies
assemble.standalone.mk to generate the PNOR properly based on the new
environment variable

Change-Id: Ic565d09626304a8159933feb1f4f0784deadbbe3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151611
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>