Fix misleading error in patch-bmc script
When the script doesn't detect standalone commit hash and the user sets
HB_PATCH_BMC_FORCE_PATCH the function that makes git patch files will
give up and not create those files but later will attempt to copy these
non-existent files to the system causing an error that terminates the
script. This leads users to believe the script failed altogether when it
did not.

This commit also updates the change-id it looks for when determining if
the standalone patch commit is present.

Change-Id: Ia02310a549cf351b2093b48e048132de588c7822
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152635
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>