ebmc setup script: update the customrc with required exports
Adds a new feature to the setup-standalone-repo script which asks the
user if they want their customrc updated to include the required
CONFIG_FILE and BMC_STANDALONE_PATCHING vars for standalone patch
building. This will allow the user to exit/enter the workon for that
repo without having to reset those vars.

The script will create a backup customrc if the user accepts changes to
their existing customrc named customrc.backup.

Change-Id: I1b6f774ca3227ba56e49e4399480d1271048cac3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152181
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>