Base dimm targets' FAPI_POS attributes on the system-wide count
This commit fixes an issue where currently the FAPI_POS attribute
for the dimm targets is based on the number of dimms per node such
that the values repeat across multi-node Denali systems.

This change sets the FAPI_POS attribute to be based on the system-wide
count for the dimm targets; thereby assuring unique values across
the whole system.

Change-Id: I6d1a5133c4fdafc9bb439ef0a14c25a684624eb1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152333
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>