Odyssey: Attr Sync - Push
This commit adds logic to push attribute blob out to Odyssey.  An
attribute blob with the current attribute values is generated by a HWP
and is pushed into Odyssey using a chipop, and a response from Odyssey
is analyzed by another HWP. This happens on all functional Odyssey
chips in the system in istep 12.2.

Change-Id: I730dbffa8730f2ad3f391790e8c0e3bbdaa89787
JIRA: PFHB-490
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152712
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>