Custom linker: Do not error on duplicate weak symbols
This commit makes a change so that the custom linker does not throw an
error when it encounters duplicate weak symbols.

This background is taken from comments in linker.C:

BACKGROUND
==========

This error path happens when a function marked "inline"
contains a static variable, and that function is compiled
into two or more shared libraries, and the compiler
elects not to inline the function in at least one spot in
two or more of those libraries. This generates symbols
marked "WEAK" (or in GCC's case, "UNIQUE", which is a
GNU-specific extension) to be merged with the other
shared libraries of a program whenever the shared library
is dynamically loaded (or in our case, when we pre-link
our libraries).

The reason this is needed is because inline functions are
defined to be "the same function" even though they can
violate the One-Definition Rule (ODR) and appear in
multiple translation units (e.g. via the containing
header file being #include'd in multiple source
files). That means that the static local variables
contained in two copies of the "same" inline function
must refer to the same variable. If this requirement is
not held, then we could call the "same" function from two
different shared objects, and they would be accessing
*different* static local variables, which could lead to
bugs.

(The reason why two libraries must have non-inline
versions of the "inline" functions generated to trigger
this error case is because the symbols are discarded when
references to them are inlined, because they generate no
relocations (the relocations already having been done in
the .so at compile-time, not at load-time). The bug is
still technically present when that happens, and this
actually happens for our Singleton<T> class, whose
instance() member function is always chosen for inlining
by the compiler (unless compiling with optimization level
-O0, which is why we can't do that). However, since we
route all accesses through a named non-inline function
API in the *same* shared object, we bypass the issue.)

NOW
===

We are now ignoring this issue because (1) it is present even when we
do not detect thie issue, and (2) this causes errors for us when it
shouldn't. We often don't care that multiple versions of the same
inline function exist in the binary. We are also aware that if static
inline functions access local static variables, there will be
issues. This is why we have "getTheInstance()" functions for our
singleton classes.

Change-Id: Ia7257497e9a67b45746e2c646d205ff3874d3ca7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152345
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>