Dynamically build IPZ formatted VPD for HDAT
HDAT used a templated binary file to build IPZ formatted VPD for RAM
areas and MS areas. This was not flexible for size changes over time.
While Odyssey still fits the templated structure, it only just does.

This commit removes that logic in favor of building up the VPD
record-by-record and keyword-by-keyword to achieve the same result but
does so without forcing size restrictions based on the templated file.

Change-Id: Ie7b6d29ac11391fae8705a0e05535477ebb018f2
JIRA: PFHB-387
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153101
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>