Disable Resource Recovery
Resource Recovery is a feature where predictive guard records
are ignored if we pass the threshold for minimum hardware to boot.

Typically this means that too many cores or dimms were guarded out.

Starting in FW1050 we will treat all guard records the same way we
treat fatal guards currently, i.e. all guards are applied and if we
do not have enough hardware to boot, we will simply fail to boot.

Change-Id: I8d87102c6b53b0a80bd22e1cccd03c7073329f76
JIRA: PFHB-474
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152779
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>