HDAT SLCA: Do not add multiple logical dimms with same PDR instance
Prior to Odyssey, one logical dimm mapped to a single FRU. This is not
the case anymore. To prevent duplicate SLCA entries for the same PDR
entity instance this commit adds logic to first check if there is
already an existing SLCA entry for the dimm PDR entity instance.

Change-Id: I1eda5dfaa2b106143cbabd3dab1cf0a4c637f118
JIRA: PFHB-387
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153198
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>