Clear SPI status register's RDR full bit
Clears the RDR full bit in p10_spi_clear_status_errors function.

A defect was opened where the RDR full bit was set in the SPI
pre-checking function. The SPI engine init code did not proactively
clear out this bit as part of engine init, but it arguably should to
make sure the engine state is pristine before hostboot attempts to use
it.  It's not obvious why the RDR would be full before Hostboot uses it
for the first time.

The FSP is not supposed to use Hostboot's engine.  The SBE is a more
likely source since it reads its code image using the SPI engine,
however if it were at fault, it would imply SBE SPI read op was
interrupted (asked for receive data but never read it).

As a proactive measure, Hostboot will just clear out the RDR full condition
when it inits the spi engine to hopefully avoid this in the future.

Change-Id: Ia357d5a80b2717a0fa541f55ce9ef2ebd8f12800
STGD:547290
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153187
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153202
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>